+++ apps/backend/src/app/api/latest/internal/projects/current/crud.tsx
@@ -12,7 +12,7 @@ export const projectsCrudHandlers = createLazyProxy(() => createCrudHandlers(pro
   onUpdate: async ({ auth, data }) => {
     if (
       data.config?.email_theme &&
-      !Object.keys(auth.tenancy.completeConfig.emails.themeList).includes(data.config.email_theme)
+      !Object.keys(auth.tenancy.completeConfig.emails.themes).includes(data.config.email_theme)
     ) {
       throw new StatusError(400, "Invalid email theme");
     }
@@ -41,8 +41,10 @@ export const projectsCrudHandlers = createLazyProxy(() => createCrudHandlers(pro
       }
     });
 
+    const prisma = await getPrismaClientForTenancy(auth.tenancy);
+
     // delete managed ids from users
-    const users = await getPrismaClientForTenancy(auth.tenancy).projectUser.findMany({
+    const users = await prisma.projectUser.findMany({
       where: {
         mirroredProjectId: 'internal',
         serverMetadata: {
@@ -57,7 +59,7 @@ export const projectsCrudHandlers = createLazyProxy(() => createCrudHandlers(pro
           (id: any) => id !== auth.project.id
         ) as string[];
 
-      await getPrismaClientForTenancy(auth.tenancy).projectUser.update({
+      await prisma.projectUser.update({
         where: {
           mirroredProjectId_mirroredBranchId_projectUserId: {
             mirroredProjectId: 'internal',