+++ apps/backend/src/lib/projects.tsx
@@ -184,7 +184,7 @@ export async function createOrUpdateProject(
         senderName: dataOptions.email_config.sender_name,
         senderEmail: dataOptions.email_config.sender_email,
       } satisfies OrganizationRenderedConfig['emails']['server'] : undefined,
-      'emails.theme': dataOptions.email_theme,
+      'emails.selectedThemeId': dataOptions.email_theme,
       // ======================= rbac =======================
       'rbac.defaultPermissions.teamMember': translateDefaultPermissions(dataOptions.team_member_default_permissions),
       'rbac.defaultPermissions.teamCreator': translateDefaultPermissions(dataOptions.team_creator_default_permissions),
@@ -232,7 +232,7 @@ export async function createOrUpdateProject(
 
   // Update owner metadata
   const internalEnvironmentConfig = await rawQuery(globalPrismaClient, getRenderedEnvironmentConfigQuery({ projectId: "internal", branchId: DEFAULT_BRANCH_ID }));
-  const prisma = getPrismaClientForSourceOfTruth(internalEnvironmentConfig.sourceOfTruth, DEFAULT_BRANCH_ID);
+  const prisma = await getPrismaClientForSourceOfTruth(internalEnvironmentConfig.sourceOfTruth, DEFAULT_BRANCH_ID);
   await prisma.$transaction(async (tx) => {
     for (const userId of options.ownerIds ?? []) {
       const projectUserTx = await tx.projectUser.findUnique({