+++ apps/e2e/tests/backend/endpoints/api/v1/auth/password/send-reset-code.test.ts
@@ -30,7 +30,7 @@ it("should send a password reset code per e-mail", async ({ expect }) => {
       },
       MailboxMessage {
         "from": "Stack Dashboard <noreply@example.com>",
-        "subject": "Mock subject, <Subject value={\`Reset your password at \${projectDisplayName}\`} />",
+        "subject": "Mock subject, <Subject value={\\"Reset your password at \\" + projectDisplayName} />",
         "to": ["<default-mailbox--<stripped UUID>@stack-generated.example.com>"],
         <some fields may have been hidden>,
       },