+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/js-example
 
+## 2.8.25
+
+### Patch Changes
+
+- @stackframe/js@2.8.25
+
+## 2.8.24
+
+### Patch Changes
+
+- @stackframe/js@2.8.24
+
 ## 2.8.23
 
 ### Patch Changes